// src/model/deck.ts
export default class Deck {
  id = "";
  name = "";
  cards: string[] = [];
  houses: string[] = [];

  // "expansion": 341,
  // "power_level": 0,
  // "chains": 0,
  // "wins": 0,
  // "losses": 0,
  // "is_my_deck": false,
  // "notes": [],
  // "is_my_favorite": false,
  // "is_on_my_watchlist": false,
  // "casual_wins": 0,
  // "casual_losses": 0,
  // "_links": {
  //     "houses": [
  //         "Dis",
  //         "Logos",
  //         "Shadows"
  //     ]
}